Refactoring.
Separate callback definition.
